+++ SenderIntBuffer.java
@@ -101,9 +101,9 @@ public class SenderIntBuffer extends Buffer{
         if (newSendBase > sendMax){
             return -1;
         }
-        AtomicInteger oldSendBase = this.sendBase;
+        int oldSendBase = this.sendBase.intValue();
         this.sendBase.set(newSendBase);
-        return oldSendBase.intValue(); 
+        return oldSendBase; 
     }
 
     public int reset(){
